refactor: 불필요한 인터페이스 계층 제거
